% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{censor_low_squish_high}
\alias{censor_low_squish_high}
\alias{squish_low_censor_high}
\title{Out of bounds handling}
\usage{
censor_low_squish_high(x, range = c(0, 1), only.finite = TRUE)

squish_low_censor_high(x, range = c(0, 1), only.finite = TRUE)
}
\arguments{
\item{x}{A numeric vector of values to modify.}

\item{range}{A numeric vector of length two giving the minimum and maximum
limit of the desired output range respectively.}

\item{only.finite}{A logical of length one. When \code{TRUE}, only finite values
are altered. When \code{FALSE}, also infinite values are altered.}
}
\description{
These functions should mostly be used for handling out of bounds data
for colour scales fo \code{\link{geom_georaster()}}. Censored values outside
the limits are removed and squished values outside the limits are set to the
maxumum limits. The functions' main purpose is to be passed to the \code{oob}
argument in a ggplot scale function, and would be particularly useful for
plotting fields such as precipitation.
}
\examples{
# Create a 2d array
z <- array(dim = c(100, 100))
for (i in 1:100) {
  for (j in 1:100) {
    z[i, j] <- sin(i / 10) + sin(j / 10)
  }
}

# Define a domain
dom <- define_domain(11, 59, 100, 0.2, "longlat")

df <- data.frame(valid_dttm = as_dttm(2024091700))
df$data <- geolist(geofield(z, domain = dom))
df <- as_harp_df(df)

# Normal plot
plot(df)

# Add scales to only show values between -1 and 1
# All values are censored
plot(df) + scale_fill_viridis_c(limits = c(-1, 1))

# Censor the low values and squish the high values
plot(df) +
  scale_fill_viridis_c(limits = c(-1, 1), oob = censor_low_squish_high)

# Censor high values and squish low values
plot(df) +
  scale_fill_viridis_c(limits = c(-1, 1), oob = squish_low_censor_high)
}
