% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_functions.R
\name{colour_scales}
\alias{colour_scales}
\alias{scale_fill_diff}
\alias{scale_fill_precip}
\alias{scale_fill_precip_b}
\title{Special colour scales}
\usage{
scale_fill_diff(
  name = "difference",
  ...,
  limits = abs_range,
  low = scales::muted("red"),
  mid = "white",
  high = scales::muted("blue"),
  direction = c(1, -1)
)

scale_fill_precip(
  ...,
  name = "mm",
  transform = "log2",
  colours = viridisLite::viridis(256, option = "G", direction = -1),
  limits = c(0.1, 32),
  oob = censor_low_squish_high,
  na.value = NA
)

scale_fill_precip_b(
  ...,
  name = "mm",
  transform = "log2",
  colours = viridisLite::viridis(256, option = "G", direction = -1),
  limits = c(0.1, 32),
  oob = censor_low_squish_high,
  na.value = NA,
  n.breaks = 8
)
}
\arguments{
\item{name}{The name of the scale. Used as the axis or legend title. If
\code{waiver()}, the default, the name of the scale is taken from the first
mapping used for that aesthetic. If \code{NULL}, the legend title will be
omitted.}

\item{...}{
  Arguments passed on to \code{\link[ggplot2:continuous_scale]{continuous_scale}}
  \describe{
    \item{\code{scale_name}}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} The name of the scale
that should be used for error messages associated with this scale.}
    \item{\code{breaks}}{One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{waiver()} for the default breaks computed by the
\link[scales:new_transform]{transformation object}
\item A numeric vector of positions
\item A function that takes the limits as input and returns breaks
as output (e.g., a function returned by \code{\link[scales:breaks_extended]{scales::extended_breaks()}}).
Note that for position scales, limits are provided after scale expansion.
Also accepts rlang \link[rlang:as_function]{lambda} function notation.
}}
    \item{\code{minor_breaks}}{One of:
\itemize{
\item \code{NULL} for no minor breaks
\item \code{waiver()} for the default breaks (none for discrete, one minor break
between each major break for continuous)
\item A numeric vector of positions
\item A function that given the limits returns a vector of minor breaks. Also
accepts rlang \link[rlang:as_function]{lambda} function notation. When
the function has two arguments, it will be given the limits and major
break positions.
}}
    \item{\code{n.breaks}}{An integer guiding the number of major breaks. The algorithm
may choose a slightly different number to ensure nice break labels. Will
only have an effect if \code{breaks = waiver()}. Use \code{NULL} to use the default
number of breaks given by the transformation.}
    \item{\code{labels}}{One of the options below. Please note that when \code{labels} is a
vector, it is highly recommended to also set the \code{breaks} argument as a
vector to protect against unintended mismatches.
\itemize{
\item \code{NULL} for no labels
\item \code{waiver()} for the default labels computed by the
transformation object
\item A character vector giving labels (must be same length as \code{breaks})
\item An expression vector (must be the same length as breaks). See ?plotmath for details.
\item A function that takes the breaks as input and returns labels
as output. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
}}
    \item{\code{limits}}{One of:
\itemize{
\item \code{NULL} to use the default scale range
\item A numeric vector of length two providing limits of the scale.
Use \code{NA} to refer to the existing minimum or maximum
\item A function that accepts the existing (automatic) limits and returns
new limits. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
Note that setting limits on positional scales will \strong{remove} data outside of the limits.
If the purpose is to zoom, use the limit argument in the coordinate system
(see \code{\link[ggplot2:coord_cartesian]{coord_cartesian()}}).
}}
    \item{\code{rescaler}}{A function used to scale the input values to the
range [0, 1]. This is always \code{\link[scales:rescale]{scales::rescale()}}, except for
diverging and n colour gradients (i.e., \code{\link[ggplot2:scale_colour_gradient2]{scale_colour_gradient2()}},
\code{\link[ggplot2:scale_colour_gradientn]{scale_colour_gradientn()}}). The \code{rescaler} is ignored by position
scales, which always use \code{\link[scales:rescale]{scales::rescale()}}. Also accepts rlang
\link[rlang:as_function]{lambda} function notation.}
    \item{\code{oob}}{One of:
\itemize{
\item Function that handles limits outside of the scale limits
(out of bounds). Also accepts rlang \link[rlang:as_function]{lambda}
function notation.
\item The default (\code{\link[scales:oob]{scales::censor()}}) replaces out of
bounds values with \code{NA}.
\item \code{\link[scales:oob]{scales::squish()}} for squishing out of bounds values into range.
\item \code{\link[scales:oob]{scales::squish_infinite()}} for squishing infinite values into range.
}}
    \item{\code{trans}}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Deprecated in favour of
\code{transform}.}
    \item{\code{call}}{The \code{call} used to construct the scale for reporting messages.}
    \item{\code{super}}{The super class to use for the constructed scale}
  }}

\item{low, high}{Colours for low and high ends of the gradient.}

\item{mid}{colour for mid point}

\item{direction}{Either \code{1} or \code{-1}. If \code{-1} the palette will be reversed.}

\item{transform}{For continuous scales, the name of a transformation object
or the object itself. Built-in transformations include "asn", "atanh",
"boxcox", "date", "exp", "hms", "identity", "log", "log10", "log1p", "log2",
"logit", "modulus", "probability", "probit", "pseudo_log", "reciprocal",
"reverse", "sqrt" and "time".

A transformation object bundles together a transform, its inverse,
and methods for generating breaks and labels. Transformation objects
are defined in the scales package, and are called \verb{transform_<name>}. If
transformations require arguments, you can call them from the scales
package, e.g. \code{\link[scales:transform_boxcox]{scales::transform_boxcox(p = 2)}}.
You can create your own transformation with \code{\link[scales:new_transform]{scales::new_transform()}}.}

\item{na.value}{Colour to use for missing values}
}
\description{
These colour scales modify \emph{ggplot} colour scales for specific purposes.
\code{scale_fill_diff} modifies \code{\link[ggplot2]{scale_fill_gradient2}} by
making the colour scale even both sides of zero, rather than the range each
side of zero. \code{scale_fill_precip} modifies
\code{\link[ggplot2]{scale_fill_gradientn}}, creating a continuous colour
scale that is suitable for showing precipitation on a map - it uses a log2
transformation such that the colours scale with powers of 2. It also sets
a minimum limit of 0.1 and squishes an upper limit of 32,
\code{scale_fill_precip2_b} is the binned variant of this scale and modifies
\code{\link[ggplot2]{scale_fill_stepsn}}.
}
\examples{
# scale_fill_diff has an even colour bar
df <- expand.grid(x = 1:100, y = 1:100)
df$value <- rnorm(nrow(df), -2)

p <- ggplot(df, aes(x, y, fill = value)) +
  geom_raster()

p + scale_fill_diff()

# Compare with legend for scale_fill_gradient2()
p + scale_fill_gradient2()
\dontrun{
  # Read in some precipitation data from the Norwegian Meteorological
  # Institute Thredds server
  library(harpIO)
  precip <- read_grid(
    "https://thredds.met.no/thredds/dodsC/metpparchive/2024/07/24/met_analysis_1_0km_nordic_20240724T15Z.nc",
    "precipitation_amount",
    file_format      = "netcdf",
    file_format_opts = netcdf_opts(proj4_var = "projection_lcc"),
    data_frame       = TRUE
  )

  p <- plot(precip)
  p

  # Use scale_fill_precip()
  p + scale_fill_precip()

  # With the banded variant
  p + scale_fill_precip_b()
}
}
