% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord-geo.R
\name{coord_geo}
\alias{coord_geo}
\title{Title}
\usage{
coord_geo(
  xlim = NULL,
  ylim = NULL,
  expand = TRUE,
  crs = NULL,
  default_crs = sf::st_crs(4326),
  datum = sf::st_crs(4326),
  label_graticule = waiver(),
  label_axes = waiver(),
  lims_method = c("cross", "box", "orthogonal", "geometry_bbox"),
  ndiscr = 100,
  default = FALSE,
  clip = "on"
)
}
\arguments{
\item{clip}{}
}
\description{
Title
}
