% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_point_verif.R
\name{dashboard_point_verifUI}
\alias{dashboard_point_verifUI}
\alias{dashboard_point_verif}
\title{Shiny module for point verification dashboard}
\usage{
dashboard_point_verifUI(id)

dashboard_point_verif(
  input,
  output,
  session,
  verif_data,
  colour_table,
  time_axis
)
}
\arguments{
\item{id}{An ID string for the module's UI}

\item{input}{input reactive}

\item{output}{output reactive}

\item{session}{Session from which to make a child scope (the default should
almost always be used).}

\item{verif_data}{verification data as a reactive value}

\item{colour_table}{A reactive data frame in a format suitable for the
\code{colour_table} argument to \code{\link{plot_point_verif}}}

\item{time_axis}{A reactive string giving the name of the time axis to
use as the x-axis in the dashboard. Mut be one of "lead_time",
"valid_dttm", or "valid_hour" and the column must exist in \code{verif_data}.}
}
\description{
Shiny module for point verification dashboard
}
\examples{

# Run examples link doesn't work for shiny apps in RStudio. Copy-paste
# example to console to run.

library(shiny)

# Set the theme to white
shinyOptions(theme = "white")

ui <- fluidPage(
  fluidRow(
    column(12, dashboard_point_verifUI("dshbrd"))
  )
)

server <- function(input, output, session) {
  # Data could have "fcst_model" or "mname" and "leadtime" or "lead_time" as
  # column names depending on the harp version
  fcst_model_col <- intersect(
    c("fcst_model", "mname"), colnames(ens_verif_data$ens_summary_scores)
  )
  lt_col <- intersect(
    c("lead_time", "leadtime"), colnames(ens_verif_data$ens_summary_scores)
  )
  col_tbl <- data.frame(
    fcst_model = unique(ens_verif_data$ens_summary_scores[[fcst_model_col]]),
    colour     = c("red", "green", "blue")
  )
  callModule(
    dashboard_point_verif, "dshbrd", reactive(ens_verif_data),
    reactive(col_tbl), reactive(lt_col)
  )
}

if (interactive()) {
  shinyApp(ui, server)
}
}
