% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{det_verif_data}
\alias{det_verif_data}
\title{Deterministic verification scores example data.}
\format{
A list with 2 data frames: \describe{
\item{det_summary_scores}{Summary scores for the forecasts}
\item{det_threshold_scores}{Categorical scores for the forecasts based on
threshold exceedence categories.}}
}
\source{
MetCoOp / HIRLAM
}
\usage{
det_verif_data
}
\description{
Example verification for 2 deterministic models: The control member of
operational MEPS and operational Arome Arctic. The scores are based on
stations that are common to the two domains for September 2018.
}
\keyword{datasets}
