% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_verif_plot.R
\name{download_verif_plotUI}
\alias{download_verif_plotUI}
\alias{download_verif_plot}
\title{Download module for point verification plots in a Shiny app}
\usage{
download_verif_plotUI(id)

download_verif_plot(
  input,
  output,
  session,
  verif_data,
  score_options,
  colour_table
)
}
\arguments{
\item{id}{An ID string for the module's UI}

\item{input}{input reactive}

\item{output}{output reactive}

\item{session}{Session from which to make a child scope (the default should
almost always be used).}

\item{verif_data}{verification data as a reactive value}

\item{score_options}{A reactive list of options for the plot. This list is
output by \code{\link{interactive_point_verif}}.}

\item{colour_table}{A reactive data frame in a format suitable for the
\code{colour_table} argument to \code{\link{plot_point_verif}}}
}
\description{
The module's UI is a "Save" button that opens a modal where the user can
select some options related to the download plot. When passing a list of
options, the following elements must be included:
\itemize{
\item "score"     - The score to plot.Must be preceded by the name of the element
in the verification list, e.g. "ens_summary_scores_mean_bias".
\item "num_cases" - TRUE/FALSE - whether to include a number of cases panel.
\item "to_y_zero" - TRUE/FALSE - whether to include 0 on the y axis.
\item "x_axis"    - the variable to plot on the x axis.
\item "facets"    - the variables to facet by - should be unquoted and wrapped in
\code{\link[ggplot2]{vars}}. Set to NULL for no faceting.
\item "filters"   - filtering expressions - should be unquoted and wrapped in
\code{\link[ggplot2]{vars}}. Set to NULL for no filtering.
\item "line_cols" - The variable to set the line colour aesthetic.
\item "highlight" - For deterministic verification of ensemble members - the
member to highlight.
\item "flip_axes" - TRUE/FALSE - whether to flip the x and y axes.
\item n_cases_pos - The position of the number of cases panel.
}
}
\examples{
library(shiny)

ui <- fluidPage(
  download_verif_plotUI("dwnld")
)

server <- function(input, output, session) {
  # Set options
  opts <- list(
    score       = "ens_summary_scores_mean_bias",
    num_cases   = FALSE,
    to_y_zero   = FALSE,
    x_axis      = "lead_time",
    facets      = NULL,
    filters     = NULL,
    line_cols   = "fcst_model",
    highlight   = NULL,
    flip_axes   = FALSE,
    n_cases_pos = "bottom"
  )

  fcst_model_col <- intersect(
    c("fcst_model", "mname"), colnames(ens_verif_data$ens_summary_scores)
  )
  col_tbl <- data.frame(
    fcst_model = unique(ens_verif_data$ens_summary_scores[[fcst_model_col]]),
    colour     = c("red", "green", "blue")
  )

  callModule(
    download_verif_plot, "dwnld", reactive(ens_verif_data),
    reactive(opts), reactive(col_tbl)
  )
}
}
