% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ens_verif_data}
\alias{ens_verif_data}
\title{Ensemble verification scores example data.}
\format{A list with 2 data frames: \describe{
  \item{ens_summary_scores}{Summary scores for the ensembles.}
  \item{ens_threshold_scores}{Categorical scores for the ensembles based on
  threshold exceedence categories.}}}
\source{
MetCoOp / HIRLAM
}
\usage{
ens_verif_data
}
\description{
Example verification for 3 ensembles: MEPS operational and two experimental
MEPS ensembles - one with and one without surface perturbations.
}
\keyword{datasets}
