% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-geocontour.R
\name{geom_geocontour}
\alias{geom_geocontour}
\title{Title}
\usage{
geom_geocontour(
  mapping = NULL,
  data = NULL,
  position = "identity",
  ...,
  upscale_factor = 1,
  upscale_method = "mean",
  downsample_location = "bottom_left",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{inherit.aes}{}
}
\description{
Title
}
