% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-geowindvec.R
\name{geom_geowindvec}
\alias{geom_geowindvec}
\title{Plot 2d wind vectors}
\usage{
geom_geowindvec(
  mapping = NULL,
  data = NULL,
  position = "identity",
  ...,
  rotate = FALSE,
  skip = 0,
  max_vec = 10,
  skip_from = "centre",
  arrow = ggplot2::arrow(length = unit(0.15, "cm")),
  arrow.fill = NULL,
  lineend = "butt",
  linejoin = "round",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}'s \code{params} argument. These
arguments broadly fall into one of 4 categories below. Notably, further
arguments to the \code{position} argument, or aesthetics that are required
can \emph{not} be passed through \code{...}. Unknown arguments that are not part
of the 4 categories below are ignored.
\itemize{
\item Static aesthetics that are not mapped to a scale, but are at a fixed
value and apply to the layer as a whole. For example, \code{colour = "red"}
or \code{linewidth = 3}. The geom's documentation has an \strong{Aesthetics}
section that lists the available options. The 'required' aesthetics
cannot be passed on to the \code{params}. Please note that while passing
unmapped aesthetics as vectors is technically possible, the order and
required length is not guaranteed to be parallel to the input data.
\item When constructing a layer using
a \verb{stat_*()} function, the \code{...} argument can be used to pass on
parameters to the \code{geom} part of the layer. An example of this is
\code{stat_density(geom = "area", outline.type = "both")}. The geom's
documentation lists which parameters it can accept.
\item Inversely, when constructing a layer using a
\verb{geom_*()} function, the \code{...} argument can be used to pass on parameters
to the \code{stat} part of the layer. An example of this is
\code{geom_area(stat = "density", adjust = 0.5)}. The stat's documentation
lists which parameters it can accept.
\item The \code{key_glyph} argument of \code{\link[ggplot2:layer]{layer()}} may also be passed on through
\code{...}. This can be one of the functions described as
\link[ggplot2:draw_key]{key glyphs}, to change the display of the layer in the legend.
}}

\item{rotate}{Whether to rotate the U and V wind components from grid
relative directions to earth relative directions prior to plotting. The
default, \code{FALSE}, is to retain vectors that are relative to the grid.}

\item{skip}{The number of grid squares to skip between each vector.}

\item{max_vec}{The magnitude of a vector with a length equal to \code{skip + 1}
grid squares.}

\item{skip_from}{The location of the grid square within a neighbourhood of
skipped grid squares.}

\item{arrow}{specification for arrow heads, as created by \code{\link[grid:arrow]{grid::arrow()}}.}

\item{arrow.fill}{fill colour to use for the arrow head (if closed). \code{NULL}
means use \code{colour} aesthetic.}

\item{lineend}{Line end style (round, butt, square).}

\item{linejoin}{Line join style (round, mitre, bevel).}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\description{
This function visualises \code{geofield} \emph{U} and \emph{V} wind data as vectors. It is
built on top of \code{\link[ggplot2]{geom_segment}}.

It requires data frames with columns containing
\code{\link[harpCore]{geofield}}s for both the u and v winds. This would
normally be a \code{\link[harpCore]{geolist}} column, but a standard list
column containing geofields will also work.

Only one row of a data frame can be plotted without faceting.
}
\details{
Wind vectors can be plotted for every grid location in a \code{geofield}. However,
this often leads to vectors that are too close to one another to be seen
clearly in the plot. Therefore, you have the option to skip some vectors.
Currently only the same skipping can be applied in the x and y directions.
The length of the vectors are set so that the magnitude set in \code{max_vec}
results in a vector that spans \code{skip + 1} grid squares. That is to say that
if \code{max_vec = 10} and \code{skip = 0}, a vector with a magnitude of 10 will be
one grid square long. The lengths of all other vectors are calculated
relative to this.
}
\section{Aesthetics}{

\code{geom_geowindvec()} understands the following aesthetics (required aesthetics are
in bold):
\itemize{
\item \code{\strong{u}}
\item \code{\strong{v}}
\item \code{\link[ggplot2:aes_colour_fill_alpha]{alpha}}
\item \code{\link[ggplot2:aes_colour_fill_alpha]{colour}}
\item \code{\link[ggplot2:aes_group_order]{group}}
\item \code{\link[ggplot2:aes_linetype_size_shape]{linetype}}
\item \code{\link[ggplot2:aes_linetype_size_shape]{linewidth}}
}
}

