% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-geoxs.R
\name{geom_xs_pressure}
\alias{geom_xs_pressure}
\title{2d vertical cross section}
\usage{
geom_xs_pressure(
  mapping = NULL,
  data = NULL,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  vert_res = 100,
  out_of_bounds = "squish",
  level_defs = harpCore::harmonie_levels$full65,
  scale_pressure = 1,
  scale_distance = 1,
  pressure_lims = c(NA, NA),
  interpolate = FALSE,
  hjust = 0.5,
  vjust = 0.5,
  topo_fill = "grey20",
  topo_colour = NA,
  topo_linewidth = 1,
  ...
)
}
\description{
These geoms visualise vertical cross sections as read in by harpIO
functions with \code{transformation = "xsection"}. For a full represenation of the
atmsophere, the data should be on vertical levels direct from the model.
Vertical interpolation to pressure levels or height levels is handled by the
geom.

\itemize{
\item{\code{geom_xs()}}{Cross section on native levels - i.e. no interpolation}
\item{\code{geom_xs_pressure()}}{Cross section on model hybrid levels interpolated to pressure levels}
\item{\code{geom_xs_height()}}{Cross section on model hybrid levels interpolated to height levels}
}
}
\section{Aesthetics}{
 \code{geom_xs()} understands the following aesthetics
(required aesthetics are in bold):
\itemize{
\item \strong{xs} - the column with the cross section data
\item alpha
}

\code{geom_xs_pressure()} understands the following aesthetics (required
aesthetics are in bold):
\itemize{
\item \strong{xs} - the column with the cross section data
\item \strong{psfc} - the column with surface pressure for the cross section
\item alpha
}

\code{geom_xs_height()} understands the following aesthetics (required
aesthetics are in bold):
\itemize{
\item \strong{xs} - the column with the cross section data
\item \strong{psfc} - the column with surface pressure for the cross section
\item \strong{temp} - the column with the cross section of air temperature
\item \strong{spec_hum} - the column with the cross section of specific humidity of air
\item \strong{topo} - the column a representation of the topography cross section
\item alpha
}
}

