% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map.R
\name{get_domain}
\alias{get_domain}
\title{Get domain information from a harp object}
\usage{
get_domain(x, ...)
}
\arguments{
\item{x}{A harp object - can be a geofield, geolist, harp spatial data frame,
harp analysis or harp forecast}

\item{col}{The column from which to extract the domain for harp objects that
include data frames}
}
\value{
An object (or list of objects) of class "geodomain"
}
\description{
Get domain information from a harp object
}
