% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map.R
\name{get_map}
\alias{get_map}
\title{Get map data for a given geodomain}
\usage{
get_map(dom = NULL, map = "world", polygon = TRUE, ...)
}
\arguments{
\item{dom}{A \code{geodomain}, \code{geofield}, \code{geolist}, or data frame with \code{geolist}
columns.}

\item{map}{The map database from which to get the map data. The default is
\code{\link[maps]{world}}. Datasets from the \code{rnaturalearth}
package may also be used, but only paths will be returned.}

\item{polygon}{Logical. Whether to return the data as polygons (\code{TRUE}), or
paths (\code{FALSE}). The default is for polygons to be returned.}

\item{...}{Not used.}

\item{col}{For a \code{harp_grid_df} data frame, the column from which to get the
domain for the map. Uses the same semantics as \code{\link[dplyr]{select}}.}
}
\value{
A data frame with columns "x", "y", and for \code{polygon = TRUE} an
additional "group" column.
}
\description{
Given a specific domain, a map is returned on that domain. The function is
intended to be used in conjunction with \code{\link{geom_geocontour}} and /
or \code{\link{geom_georaster}} when plotting using \code{ggplot()}.
}
\examples{
# Define a domain and make some data
arr <- array(dim = c(200, 200))
for (i in 1:200) {
  for (j in 1:200) {
    arr[i, j] <- sin(i / 10) + sin(j / 10)
  }
}

dom <- define_domain(10, 60, 200, 10000)
geo <- geofield(arr, domain = dom)
data <- data.frame(geo = geolist(geo))

# Use polygons
map <- get_map(geo)
ggplot() +
  geom_polygon(
    aes(x, y, group = group), map,
    fill = "seagreen", colour = "grey10"
  ) +
  geom_geocontour(aes(geofield = geo, colour = after_stat(level)), data) +
  coord_equal(expand = FALSE) +
  theme_harp_map()

# Use paths
map <- get_map(geo, polygon = FALSE)
ggplot() +
  geom_georaster(aes(geofield = geo), data) +
  geom_path(aes(x, y), map, colour = "grey30") +
  coord_equal(expand = FALSE) +
  theme_harp_map()

}
