% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_selectors.R
\name{group_selectorsUI}
\alias{group_selectorsUI}
\alias{group_selectors}
\title{Shiny module for selecting values for verification groups}
\usage{
group_selectorsUI(id)

group_selectors(input, output, session, verif_data)
}
\arguments{
\item{id}{An ID string for the module's UI}

\item{input}{input reactive}

\item{output}{output reactive}

\item{session}{Session from which to make a child scope (the default should
almost always be used).}

\item{verif_data}{verification data as a reactive value}
}
\value{
The verification data as a reactive value filtered to the values
selected for each group.
}
\description{
This module inspects the input data and identifies columns by which the
verification is grouped. A \code{\link[shiny]{selectInput}} dropdown is
created for each group and updated depending on the selections in other
groups. The verification data are filtered to the selected values for each
group.
}
\examples{
library(shiny)

ui <- fluidPage(
  fluidRow(
    group_selectorsUI("grps")
  ),
  fluidRow(
    plotOutput("plt")
  )
)

server <- function(input, output, session) {
  grp_data <- callModule(group_selectors, "grps", reactive(verif_data_grp))
  output$plt <- renderPlot({
    plot_point_verif(req(grp_data()), spread)
  })
}

if (interactive()) {
  shinyApp(ui, server)
}
}
