% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_linespline.R, R/geom_ribbonspline.R
\docType{data}
\name{GeomLinespline}
\alias{GeomLinespline}
\alias{StatLinespline}
\alias{GeomRibbonspline}
\alias{StatRibbonspline}
\title{Geom Proto}
\description{
Geom Proto

Geom Proto
}
\keyword{datasets}
\keyword{internal}
