% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harpVis-package.R
\docType{package}
\name{harpVis-package}
\alias{harpVis}
\alias{harpVis-package}
\title{harpVis: Visualisation functions for harp.}
\description{
Functions for visualising data such as verification scores, model fields, profiles, point forecasts etc. produced by harp. Each visualisation has a corresponding interactive shiny app.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/harphub/harpVis}
  \item Report bugs at \url{https://github.com/harphub/harpVis/issues}
}

}
\author{
\strong{Maintainer}: Andrew Singleton \email{andrewts@met.no}

}
\keyword{internal}
