% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_eps.R
\name{interactive_eps}
\alias{interactive_eps}
\title{Title}
\usage{
interactive_eps(input, output, session, verif_data, colour_table, bg_colour)
}
\arguments{
\item{bg_colour}{}
}
\value{

}
\description{
Title
}
