% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_basic.R
\name{plot_basic}
\alias{plot_basic}
\title{Plot "basic" scores that have only one value column,
such as mse, bias, mae etc}
\usage{
plot_basic(
  plot_data,
  score_name,
  point_size = 1.2,
  extend_y_to_zero = TRUE,
  line_width = 1,
  y_label = "",
  x_label = "Forecast length",
  flip_axes = FALSE,
  plot_num_cases = FALSE,
  ...
)
}
\arguments{
\item{score_name}{Name of the score to plot}
}
\description{
\code{plot_spatial_verif} is used to plot verification scores computed by
functions from the harpSpatial package.
}
