% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_field.R
\name{plot_domain}
\alias{plot_domain}
\alias{plot_domain.geodomain}
\alias{plot_domain.geofield}
\title{Plot a geodomain}
\usage{
plot_domain(x, ...)

\method{plot_domain}{geodomain}(x, ...)

\method{plot_domain}{geofield}(x, ...)
}
\arguments{
\item{x}{A geodomain or geofield}

\item{...}{
  Arguments passed on to \code{\link[meteogrid:plot.geodomain]{meteogrid::plot.geodomain}}
  \describe{
    \item{\code{col}}{Colour of the map contours. }
    \item{\code{mapfill}}{Colour(s) used if fill=TRUE to fill the map contours. First colour is for internal (land), second for background (sea), but this only happens if add=FALSE).}
    \item{\code{fill}}{If TRUE, the map is drawn as coloured polygons, not just lines.}
    \item{\code{interior}}{If FALSE, only the physical land boundaries are shown, not the interior political borders.}
    \item{\code{map.database}}{A character string for the map database. Default is 'world' from the \code{maps} package.}
    \item{\code{add.dx}}{If TRUE, the domain size is extended by half a grid interval in all directions, i.e. the co-ordinates are treated as the centers of grid boxes.}
    \item{\code{box}}{If TRUE, a rectangular box is drawn around the domain.}
    \item{\code{add}}{If TRUE, the plot is superposed on the current plot.}
    \item{\code{asp}}{Aspect ratio. Not considered if add=TRUE. For Lat/Lon domains, NULL may be a better choice!.}
  }}
}
\description{
Plots a map of the domain of a geodomain or geofield object. Uses the plot
method for geodomains from the meteogrid package.
}
