% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_field.R
\name{plot_field}
\alias{plot_field}
\alias{plot_field.harp_grid_df}
\alias{plot_field.harp_list}
\alias{plot_field.geofield}
\title{Plot a harp_fcst 2d field}
\usage{
plot_field(
  .fcst,
  palette = viridisLite::viridis(255),
  num_breaks = 15,
  breaks = NULL,
  legend = TRUE,
  title = "auto",
  zoom_centre = NULL,
  zoom_length = 100,
  ...
)

\method{plot_field}{harp_grid_df}(
  .fcst,
  palette = viridisLite::viridis(255),
  num_breaks = 15,
  breaks = NULL,
  legend = TRUE,
  title = "auto",
  zoom_centre = NULL,
  zoom_length = 100,
  fcst_dttm,
  lead_time,
  filter_by = NULL,
  plot_col = NULL,
  ...
)

\method{plot_field}{harp_list}(
  .fcst,
  palette = viridisLite::viridis(255),
  num_breaks = 15,
  breaks = NULL,
  legend = TRUE,
  title = "auto",
  zoom_centre = NULL,
  zoom_length = 100,
  fcst_model = NULL,
  fcst_dttm,
  lead_time,
  filter_by = NULL,
  plot_col = NULL,
  ...
)

\method{plot_field}{geofield}(
  .fcst,
  palette = viridisLite::viridis(255),
  num_breaks = 15,
  breaks = NULL,
  legend = TRUE,
  title = "auto",
  zoom_centre = NULL,
  zoom_length = 100,
  ...
)
}
\arguments{
\item{.fcst}{A harp_list or harp_grid_df data frame.}

\item{palette}{Colour palette to use. This should be a vector of colours.}

\item{num_breaks}{Number of colour breaks to use in the plot.}

\item{breaks}{The values to use for colour breaks. If not NULL, breaks has
priority over num_breaks.}

\item{legend}{Logical. Whether to plot a legend.}

\item{title}{The title of the plot. Set to "auto" to automatically generate a
title.}

\item{zoom_centre}{The centre point for a zoomed plot. Should be a 2-element
vector with the longitude and latitude of the desired centre point.}

\item{zoom_length}{A 2 element vector given the x and y lengths in number of
grid squares of the zoomed plot.}

\item{fcst_dttm}{The forecast date to plot in "YYYYMMDDhh" or similar format.
Can be omitted if there is only one date in the data.}

\item{lead_time}{The lead time to plot. Can be omitted if there is only one
lead time in the data.}

\item{filter_by}{Expressions that return a logical value wrapped inside the
\code{vars} function for filtering the data prior to plotting. Only a
single row should be left in the data frame for \code{plot_field} to work.}

\item{plot_col}{The column in the data frame to take data from to plot}

\item{fcst_model}{The fcst_model in a \code{harp_list} to plot from.}
}
\value{
A plot
}
\description{
This function is for quickly plotting a single \code{geofield}. For geofields in
\code{harp_grid_df} data frames or \code{harp_list}s, extra arguments are available for
extracting the correct \code{geofield}.
}
