% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_point_map.R
\name{plot_point_map}
\alias{plot_point_map}
\title{Plot point data on a map}
\usage{
plot_point_map(
  .data,
  col,
  map_db = "world",
  proj = NULL,
  expand = 0.1,
  polygon = FALSE,
  land_fill = "#EEEEEE",
  land_stroke = "black",
  sea_fill = "white",
  point_size = 3,
  point_stroke = "black"
)
}
\arguments{
\item{.data}{A data frame. Must include columns \code{"lon"} and \code{"lat"}.}

\item{col}{The column for which to plot points. Can be passed quoted or
unquoted. If passing a variable for the column name, it should be embraced
in double curly braces, i.e. \code{{{ }}}}

\item{map_db}{The map database to get the map from. By default this is
\code{\link[maps]{world}}, a 1:50m resolution map. You can also choose maps
from the rnaturalearth set of packages, but note that it can be problematic
to plot polygons with the \code{rnaturalearthhires::countries10} dataset.}

\item{proj}{The projection of the map. By default the data are plotted on
lon and lat coordinates. Here you can pass a proj4 string, or a \code{gemdomain},
\code{geofield} or \code{geolist} to define a domain. If a domain is passed, the
data will be cropped to that domain if any points fall outside of the
domain.}

\item{expand}{An expansion around the data. In order to prevent points from
being plotted too close to the edges of the map you can provide an
expansion factor. Can be a one or two element vector - if a one element
vector the expansion is applied in all directions, if a two element vector,
the first value is the expansion applied in the left-right direction and
the second in the bottom-top direction. By default a 10\% expansion (0.1) is
applied in all directions.}

\item{polygon}{Logical. Whether to get the map data as polygons (\code{TRUE}), or
paths (\code{FALSE}). The default is to get paths. If you experience issues
plotting polygons with \code{map_db = "world"}, try using
\code{map_db = rnaturalearthdata::countries50}.}

\item{land_fill}{If \code{polygon = TRUE}, the colour of land areas.}

\item{land_stroke}{The colour of country outlines.}

\item{sea_fill}{If \code{polygon = TRUE}, the colour of sea areas.}

\item{point_size}{The size of the data points.}

\item{point_stroke}{The outline colour of data points.}
}
\value{
A ggplot plot
}
\description{
Plots data as points on a map, with the colour scaled according to the
selected column. The output is a ggplot so can be modified by adding ggplot2
scales, coords and themes etc. By default the map is plotted using longitude
and latitude coordinates. You can change this by passing a proj4 string, or a
\code{geodomain} on which you want the data to be plotted. This function is called
by \code{\link{plot_point_verif()}} for plotting point verifcation scores on
a map.
}
\examples{
# Make some fake data
df <- subset(station_list, SID > 1000 & SID < 4000 | SID > 5000 & SID < 8500)
df$fcst <- abs(rnorm(nrow(df)))
plot_point_map(df, fcst)

# Use a mercator projection
plot_point_map(df, fcst, proj = "+proj=merc")

# Plot map as filled polygons
plot_point_map(
  df, fcst, polygon = TRUE, map_db = rnaturalearthdata::countries50,
  land_fill = scales::muted("green"),
  sea_fill  = scales::muted("blue")
)
}
