% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sal.R
\name{plot_sal}
\alias{plot_sal}
\title{Plot SAL scores}
\usage{
plot_sal(
  plot_data,
  score_name = "SAL",
  point_size = 4,
  extend_y_to_zero = TRUE,
  line_width = 1,
  quantiles = seq(0.25, 0.75, 0.05),
  xylim = "auto",
  show_stats = FALSE,
  ...
)
}
\arguments{
\item{plot_data}{A tibble from \code{plot_spatial_verif} that contains necessary scores.}

\item{score_name}{Name of the score to plot. Technically not really necessary
but makes it more intuitive when calling from \code{plot_spatial_verif}.}

\item{point_size}{The size of points to plot.}

\item{extend_y_to_zero}{Logical. Whether to extend the y-axis to include}

\item{line_width}{The width of lines to plot. The default is 1.}

\item{quantiles}{A vector of quantiles that will be used to highlight an area of S/A values within those quantiles}

\item{xylim}{Plot x/y axis shared limit. Default "auto" will generate limits based on plot data.}

\item{show_stats}{An option to show median and mean values of S, A and L as a table next to the plot. Default FALSE.}
}
\description{
Plot SAL scores
}
