% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Scatter plot of forecast vs observations}
\usage{
plot_scatter(.fcst, fcst_model, parameter, members = "all",
  facet_members = TRUE, binwidth = NULL, colours = NULL, ...)
}
\arguments{
\item{.fcst}{A \code{harp_fcst} type object.}

\item{fcst_model}{The model in the \code{harp_fcst} object to make a scatter
plot for - must not be quoted.}

\item{parameter}{The parameter to plot - used to identify the observations
column in \code{.fcst}. Must not be quoted.}

\item{members}{The member(s) to include in the forecast statistics if
\code{.fcst} is for EPS forecasts. Must be either a numeric vector of
member numbers or "all" to include all members.}

\item{binwidth}{A 2 element numeric vector giving the width of the bins to
plot for the observed and forecast directions respectively. Set to NULL
(the default) to use the deafult of the width that gives 30 bins.}

\item{colours}{A chracter vector of colours used to generate a colour
gradient.}

\item{...}{Options for \link[ggplot2]{scale_colour_gradientn}}
}
\value{
A hexbin plot.
}
\description{
Scatter plot of forecast vs observations
}
