% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scorecard.R
\name{plot_scorecard}
\alias{plot_scorecard}
\title{Title}
\usage{
plot_scorecard(
  bootstrap_data,
  fcst_model,
  ref_model,
  scores,
  facet_by = vars(score),
  num_facet_rows = 1,
  grid_facets = FALSE,
  filter_by = NULL,
  significance_breaks = c(-1.1, -0.997, -0.95, -0.68, 0.68, 0.95, 0.997, 1.1),
  colours = c("#CA0020", "#CA0020", "#CA0020", "grey70", "#0571B0", "#0571B0",
    "#0571B0"),
  fills = c("#CA0020", "#F4A582", NA, "grey70", NA, "#92C5DE", "#0571B0"),
  shapes = c(25, 25, 25, 22, 24, 24, 24),
  sizes = c(3, 2, 1, 0.5, 1, 2, 3),
  legend_labels = "auto",
  num_facet_cols = length(scores),
  ...
)
}
\arguments{
\item{num_facet_cols}{}
}
\value{

}
\description{
Title
}
