% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spatial_nact.R
\name{plot_spatial_nact}
\alias{plot_spatial_nact}
\title{Plot scores from neighborhood based contingency tables
based on Stein and Stoop article, 2019 Jan
Neighborhood-Based Contingency Tables Including Errors Compensation}
\usage{
plot_spatial_nact(
  plot_data,
  score_name = "NACT",
  point_size = 1.2,
  extend_y_to_zero = TRUE,
  line_width = 1,
  y_label = "auto",
  x_label = "auto",
  flip_axes = FALSE,
  nact_scores = list("all"),
  colour_by = "scale",
  num_facet_cols = NULL,
  ...
)
}
\arguments{
\item{plot_data}{A tibble from \code{plot_spatial_verif} that contains necessary scores.}

\item{score_name}{In this context, table name. Technically not really necessary
but makes it more intuitive when calling from \code{plot_spatial_verif}.}

\item{point_size}{The size of points to plot.}

\item{extend_y_to_zero}{Logical. Whether to extend the y-axis to include
zero.}

\item{line_width}{The width of lines to plot. The default is 1.}

\item{y_label}{Label for the y-axis. Set to "auto" to use "Score". Anything else inside
quotes will be used as the y-axis label.}

\item{x_label}{Label for the x-axis. Set to "auto" to use threshold
or scale depending on the data. Anything else inside
quotes will be used as the x-axis label.}

\item{flip_axes}{Logical of whether to swap the x and y axes.}

\item{nact_scores}{Actual scores to plot.
Available scores based on NACT are "fbias", "pod", "far", "pss", "hss", "sedi".}

\item{colour_by}{The column to colour the plot lines by.
Can be an option between scale or threshold.}

\item{num_facet_cols}{Number of columns in the faceted plot.}
}
\description{
Plot scores from neighborhood based contingency tables
based on Stein and Stoop article, 2019 Jan
Neighborhood-Based Contingency Tables Including Errors Compensation
}
