% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spatial_verif.R
\name{plot_spatial_verif}
\alias{plot_spatial_verif}
\title{Plot spatial verification scores}
\usage{
plot_spatial_verif(verif_data, score, filter_by = NULL, ...)
}
\arguments{
\item{verif_data}{Output from \link[harpSPatial]{spatial_verify}}

\item{score}{The score to plot. Currently only SAL, FSS are available.
More to come.}

\item{filter_by}{Filter the data before plotting. Must be wrapped inside the
\link[dplyr]{vars} function. This can be useful for making a single plot
where there are many groups. For example, if the data contains various models
the data can be filtered with
e.g. \code{filter_by = vars(det_model == "be13", fctime == 0)}.}
}
\description{
\code{plot_spatial_verif} is used to plot verification scores computed by
functions from the harpSpatial package.
}
