% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_station_ts.R
\name{plot_station_ts}
\alias{plot_station_ts}
\alias{plot_station_ts.harp_ens_point_df}
\title{Plot a time series for point data}
\usage{
plot_station_ts(
  .data,
  SID,
  fcst_dttm,
  x_axis = "lead_time",
  fcst_geom = "line",
  fcst_geom_args = list(),
  fcst_colour_by = NULL,
  fcst_colours = NULL,
  obs_col = NULL,
  obs_geom = "point",
  obs_geom_args = list(),
  facet_by = NULL,
  num_facet_cols = 1,
  facet_scales = "free_y",
  smooth = FALSE,
  ...
)

\method{plot_station_ts}{harp_ens_point_df}(
  .data,
  SID,
  fcst_dttm,
  x_axis = "lead_time",
  fcst_geom = "boxplot",
  fcst_geom_args = list(),
  fcst_colour_by = NULL,
  fcst_colours = NULL,
  obs_col = NULL,
  obs_geom = "point",
  obs_geom_args = list(),
  facet_by = NULL,
  num_facet_cols = 1,
  facet_scales = "free_y",
  smooth = FALSE,
  quantiles = NULL,
  best_guess = NULL,
  best_guess_geom = "line",
  best_guess_geom_args = list(),
  ...
)
}
\arguments{
\item{.data}{A \code{harp_point_df} data frame, or a \code{harp_list} containing
\code{harp_point_df} data frames.}

\item{SID}{The ID of the station(s) to plot. If more than one SID is asked
for then \code{SID} should be included in \code{facet_by}.}

\item{fcst_dttm}{The start time(s) of the the forecast to plot. If more than
one \code{fcst_dttm} is asked for, \code{fcst_dttm} should be included in \code{facet_by}.}

\item{x_axis}{The x axis of the plot.}

\item{fcst_geom}{The geom to use to plot the forecast data (see details).}

\item{fcst_geom_args}{Arguments to the \code{fcst_geom} geom as a named list.}

\item{fcst_colour_by}{Which column in \code{.data} to use to control the colours
of the forecast data.}

\item{fcst_colours}{A vector of colours to use for the forecast data. It
should be the same length as the number of colours to appear in the plot.
Where the colours are a controlled by the data, this can be a named vector
or a data frame with column names equal to the column in the data
controlling the colour and "colour".}

\item{obs_col}{If observations are to be included in the plot, the column
containing the observations data.}

\item{obs_geom}{The geom to use to plot the observations data.}

\item{obs_geom_args}{Arguments to the \code{obs_geom} geom as a named list.}

\item{facet_by}{The column(s) to use to facet the plot into panels.}

\item{num_facet_cols}{The number of columns in a faceted plot.}

\item{facet_scales}{Whether the scales should be fixed. Defaults to
\code{"free_y"}. See \code{\link[ggplot2]{facet_wrap}} for more details.}

\item{smooth}{For line and ribbon plots, whether to smooth the lines by
drawing an X-spline relative to control points. In the background
\code{\link{geom_linespline}} and \code{\link{geom_ribbonspline}} are used.}

\item{...}{Other arguments passed to methods.}

\item{quantiles}{For \code{geom = "ribbon"}, or \code{geom = "col"}, the quantiles
used to stratify the probabilities of an ensemble forecast.}

\item{best_guess}{What to plot as a "best guess" forecast. Can be any
function as a character string that reduces a vector to a single value. Can
also be an ensemble member as a numeric value or a string that is the same
as the member in a \code{harp_ens_point_df} data frame that has had the
members pivoted using \code{\link[harpCore]{pivot_members}}, e.g.
\code{"mbr000"}.}

\item{best_guess_geom}{The geom to use to plot the best guess forecast.}

\item{best_guess_geom_args}{Arguments to \code{best_guess_geom} as a named list.}
}
\value{
A ggplot object that can be saved with \code{\link[ggplot2]{ggasve}}.
}
\description{
This function plots a time series of data from a \code{harp_point_df} data frame,
or a \code{harp_list} containing \code{harp_point_df} data frames. The plotting is done
using \href{https://ggplot2.tidyverse.org/}{ggplot2} and thus uses some of
the same terminology in its arguments. Data are plotted using "geoms", and
the plots are divided into panels using facets. If the data contain a column
of observed values these may also be included in the plot, and for ensemble
data a best guess forecast may be derived from the data.
}
\section{Geoms}{


The data are plotted using geoms from ggplot2. You can control which geom
is used for each aspect of the plot using the respective arguments:
\itemize{
\item \code{fcst_geom} for forecast;
\item \code{obs_geom} for observations;
\item \code{best_guess_geom} for the "best guess" forecast from an ensemble.
}

The geom should be specified as a character string based on the geom
function from ggplot2 with the "geom_" prefix removed. For example, for a
line plot for forecast data use \code{fcst_geom = "line"}. Other arguments to
the geom function can be provided as named list to the appropriate
\verb{*_geom_args} argument to control, for example, the colour or size of the
geom.

Note that aesthetic mappings to the geom cannot be controlled, except for
\strong{x} via the \code{x_axis} argument and \strong{colour} and \strong{fill}
via the \code{colour_by} argument.
}

\section{Ensembles}{


For ensemble forecasts, some data manipulation is done prior to plotting
depending on the geom as listed below. For geoms not included below, no
manipulation is done and the plots may be difficult to interpret or not
work at all.
\itemize{
\item \strong{boxplot} - The data are grouped by the \code{x_axis} argument such
that each box is representative of the ensemble distribution for each point
on the x-axis. See \code{\link[ggplot2]{geom_boxplot}} for how the hinges,
whiskers and outliers are defined.
\item \strong{violin} - The data are grouped by the \code{x_axis} argument such
that each "violin" is representative of the ensemble distribution for each
point on the x-axis.
\item \strong{line} - The data are grouped by each ensemble member and a time
series is plotted for each member. This is the common ensemble "spaghetti"
plot.
\item \strong{ribbon} - The data are divided into bands depending on the
quantiles provided in the \code{quantiles} argument. The bands are centred such
that the the outer band is between the lowest and highest quantiles with
inner bands added until the innermost pair of quantiles is reached.
Consequently an even number of quantiles must be provided.
\item \strong{col} - The data are divided into bands of increasing quantile
pairs starting provided in the \code{quantiles} argument, with the minimum value
(quantile = 0). This gives columns of stacked probabilities staring at 0
that is particularly useful for accumulated variables such as
precipitation, or variables that truncate at 0 such as wind speed or cloud
or cloud cover.
}

A "best_guess" forecast can be added to the plot using the \code{best_guess}
argument. This can either be the name of a function that reduces the
ensemble to a single value (e.g. "mean", "median"), or the ensemble member
to treat as the best guess (e.g. 0, or "mbr000"). The geom and its options
can be provided via the \code{best_guess_geom} and \code{best_guess_geom_args}
arguments, but they must be geoms that understand the \strong{x} and
\strong{y} aesthetics.
}

\section{Filtering and faceting}{


By default, all of the data that the function is given are plotted. In many
cases this can result in overplotting. For data at more than one station
(SID), or for more than one forecast start time (fcst_dttm), there are
arguments to filter the data prior to plotting based on those values.
Otherwise filtering should be done (e.g. using \code{\link[dplyr]{filter}})
prior to passing the data to this function.

Alternatively multi-panel plots can be made using the \code{facet_by} argument.
This groups the data based on the values in the columns provided to
\code{facet_by} and plots each group of data in its own panel. The default
behaviour is to plot 1 column of panels so that the x axis lines up for all
panels, but this can be changed using the \code{num_facet_cols} argument.
Additionally, the scale of the y axis for each panel is determined by the
data for that panel, but can be common for all panels by setting
\code{facet_scales = "fixed"}.
}

\section{Observations}{


If the data include an observations column (e.g. from running
\code{\link[harpCore]{join_to_fcst}}), these observations can be included
in the plot be providing the name of the column that contains the
observations via the \code{obs_col} argument. The geom and its options can be
provided via the \code{obs_geom} and \code{obs_geom_args} arguments, but they must be
geoms that understand the \strong{x} and \strong{y} aesthetics.
}

