% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_point_map.R
\name{point_map_opts}
\alias{point_map_opts}
\title{Options for point map plotting}
\usage{
point_map_opts(
  map_db = "world",
  proj = NULL,
  expand = 0.1,
  polygon = FALSE,
  land_fill = "#EEEEEE",
  land_stroke = "black",
  sea_fill = "white",
  point_size = 3,
  point_stroke = "black"
)
}
\arguments{
\item{map_db}{The map database to get the map from. By default this is
\code{\link[maps]{world}}, a 1:50m resolution map. You can also choose maps
from the rnaturalearth set of packages, but note that it can be problematic
to plot polygons with the \code{rnaturalearthhires::countries10} dataset.}

\item{proj}{The projection of the map. By default the data are plotted on
lon and lat coordinates. Here you can pass a proj4 string, or a \code{gemdomain},
\code{geofield} or \code{geolist} to define a domain. If a domain is passed, the
data will be cropped to that domain if any points fall outside of the
domain.}

\item{expand}{An expansion around the data. In order to prevent points from
being plotted too close to the edges of the map you can provide an
expansion factor. Can be a one or two element vector - if a one element
vector the expansion is applied in all directions, if a two element vector,
the first value is the expansion applied in the left-right direction and
the second in the bottom-top direction. By default a 10\% expansion (0.1) is
applied in all directions.}

\item{polygon}{Logical. Whether to get the map data as polygons (\code{TRUE}), or
paths (\code{FALSE}). The default is to get paths. If you experience issues
plotting polygons with \code{map_db = "world"}, try using
\code{map_db = rnaturalearthdata::countries50}.}

\item{land_fill}{If \code{polygon = TRUE}, the colour of land areas.}

\item{land_stroke}{The colour of country outlines.}

\item{sea_fill}{If \code{polygon = TRUE}, the colour of sea areas.}

\item{point_size}{The size of the data points.}

\item{point_stroke}{The outline colour of data points.}
}
\description{
Generate a list of options for calling \code{\link{plot_point_map()}}. This
is useful for passing options for map plots to
\code{\link{plot_point_verif()}}.
}
