% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_lookup.R
\name{score_lookup}
\alias{score_lookup}
\title{Convert a score from a long name to a short name, or vice versa}
\usage{
score_lookup(x)
}
\arguments{
\item{x}{The name of the score to look up}
}
\value{
The short name or long name of the score
}
\description{
\code{score_lookup} looks for the input in a dictionary - if it corresponds to
a short name, the long name is returned, and if it corresponds to a long name,
the short name is returned. The argument is returned unchanged if it is not
found in the dictionary. If no argument is given, the dictionary is returned.
}
\examples{
score_lookup("SEDI")
score_lookup("mr")
score_lookup("heidke_skill_score")
}
