% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_plot_point_verif.R
\name{shiny_plot_point_verif}
\alias{shiny_plot_point_verif}
\title{Start shiny app to plot point verification}
\usage{
shiny_plot_point_verif(
  start_dir = NULL,
  full_dir_navigation = TRUE,
  theme = c("dark", "light", "white"),
  online = TRUE,
  min_cases = 1,
  ...
)
}
\arguments{
\item{start_dir}{The directory start the app in. The directory tree above
this directory will not be visible.}

\item{full_dir_navigation}{Whether to have full navigation of the directory
tree below \code{app_start_dir} via a modal (TRUE - the default), or a dropdown
selector of directories that contain harp point verification files (FALSE).}

\item{theme}{The colour theme of the app - can be one of "dark", "light" or
"white".}

\item{online}{Logical. Whether the app should have access to the internet
(e.g. to download fonts) or not.}

\item{min_cases}{The minimum number of cases required for data to be
plotted. Rows with fewer than \code{min_cases} in the \code{"num_cases"} column for
summary scores, and the \code{"num_cases_for_threshold_total"} column for
threshold scores will be removed at read time.}

\item{...}{Options to \code{\link[shiny]{runApp}}.}
}
\description{
Start shiny app to plot point verification
}
