% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_plot_spatial_verif.R
\name{shiny_plot_spatial_verif}
\alias{shiny_plot_spatial_verif}
\title{Start shiny app to plot spatial verification}
\usage{
shiny_plot_spatial_verif(start_dir = NULL, ...)
}
\description{
Start shiny app to plot spatial verification
}
