% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_common_cases.R
\name{spatial_common_cases}
\alias{spatial_common_cases}
\title{Filter spatial scores to common cases only (based on parameter, forecast
start data, and leadtime). Used only as part of \code{plot_spatial_verif}.}
\usage{
spatial_common_cases(verif_data)
}
\arguments{
\item{verif_data}{Output from \link[harpSpatial]{spatial_verify}. Must be a
data frame.}
}
\description{
Filter spatial scores to common cases only (based on parameter, forecast
start data, and leadtime). Used only as part of \code{plot_spatial_verif}.
}
