% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_plot_func.R
\name{spatial_plot_func}
\alias{spatial_plot_func}
\title{Pick appropriate functions for spatial scores}
\usage{
spatial_plot_func(score_name, ...)
}
\arguments{
\item{score_name}{A string containing the name of the score calculated from \code{verify_spatial}.
This will return the appropriate plotting function to \code{plot_spatial_verif}.}
}
\description{
Pick appropriate functions for spatial scores
}
