% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{verif_data_det}
\alias{verif_data_det}
\title{Deterministic verification scores example data.}
\format{
A \code{harp_verif} list with 2 data frames: \describe{
\item{det_summary_scores}{Summary scores for the forecasts}
\item{det_threshold_scores}{Categorical scores for the forecasts based on
threshold exceedence categories.}}
}
\source{
MetCoOp / MET Norway
}
\usage{
verif_data_det
}
\description{
Example verification for 2 deterministic models: The control member of
operational MEPS and operational Arome Arctic. The scores are based on
stations that are common to the two domains for 2 days in August 2023.
}
\keyword{datasets}
