% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{verif_data_ens}
\alias{verif_data_ens}
\title{Ensemble verification scores example data.}
\format{
A \code{harp_verif} list with 3 data frames: \describe{
\item{ens_summary_scores}{Summary scores for the ensembles.}
\item{ens_threshold_scores}{Categorical scores for the ensembles based on
threshold exceedence categories.}
\item{det_summary_scores}{Deterministic summary scores for each ensemble
member}}
}
\source{
MetCoOp
}
\usage{
verif_data_ens
}
\description{
Example verification for 2 ensembles: meps and meps_obs_error. meps is the
operational NWP ensemble for the METCOOP consortium, and meps_obs_error is
the same data, but each ensemble member is perturbed to simulate the
impact of observation errors on the verification. The scores are based on
forecasts for 2 days in August 2023.
}
\keyword{datasets}
