% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{verif_data_grp}
\alias{verif_data_grp}
\title{Grouped verification scores example data}
\format{
A \code{harp_verif} list with 1 data frame: \describe{
\item{ens_summary_scores}{Summary scores for the ensembles.}}
}
\source{
MetCoOp
}
\usage{
verif_data_grp
}
\description{
Example verification data grouped by station_group and type. The forecasts
are as \code{\link{verif_data_ens}}, but the verification is split into
groups.
}
\keyword{datasets}
