% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{verif_data_prf}
\alias{verif_data_prf}
\title{Deterministic verification scores for vertical profile example data.}
\format{
A \code{harp_verif} list with 1 data frame: \describe{
\item{det_summary_scores}{Summary scores for the forecasts}}
}
\source{
MetCoOp / MET Norway
}
\usage{
verif_data_prf
}
\description{
Example verification for 2 deterministic models: The control member of
operational MEPS and operational Arome Arctic. The scores are based on
stations and pressure levels that are common to the two domains for 2 days in
August 2023.
}
\keyword{datasets}
