% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_verif_plot.R
\name{download_verif_plot}
\alias{download_verif_plot}
\title{Title}
\usage{
download_verif_plot(
  input,
  output,
  session,
  verif_data,
  score_options,
  colour_table
)
}
\arguments{
\item{colour_table}{}
}
\description{
Title
}
