% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-geocontour.R
\name{geom_geocontour}
\alias{geom_geocontour}
\alias{geom_geocontour_filled}
\title{2d contours of a geofield}
\usage{
geom_geocontour(
  mapping = NULL,
  data = NULL,
  position = "identity",
  ...,
  bins = NULL,
  binwidth = NULL,
  breaks = NULL,
  lineend = "butt",
  linejoin = "round",
  linemitre = 10,
  upscale_factor = 1,
  upscale_method = "mean",
  downsample_location = "centre",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)

geom_geocontour_filled(
  mapping = NULL,
  data = NULL,
  position = "identity",
  ...,
  bins = NULL,
  binwidth = NULL,
  breaks = NULL,
  upscale_factor = 1,
  upscale_method = "mean",
  downsample_location = "centre",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}'s \code{params} argument. These
arguments broadly fall into one of 4 categories below. Notably, further
arguments to the \code{position} argument, or aesthetics that are required
can \emph{not} be passed through \code{...}. Unknown arguments that are not part
of the 4 categories below are ignored.
\itemize{
\item Static aesthetics that are not mapped to a scale, but are at a fixed
value and apply to the layer as a whole. For example, \code{colour = "red"}
or \code{linewidth = 3}. The geom's documentation has an \strong{Aesthetics}
section that lists the available options. The 'required' aesthetics
cannot be passed on to the \code{params}. Please note that while passing
unmapped aesthetics as vectors is technically possible, the order and
required length is not guaranteed to be parallel to the input data.
\item When constructing a layer using
a \verb{stat_*()} function, the \code{...} argument can be used to pass on
parameters to the \code{geom} part of the layer. An example of this is
\code{stat_density(geom = "area", outline.type = "both")}. The geom's
documentation lists which parameters it can accept.
\item Inversely, when constructing a layer using a
\verb{geom_*()} function, the \code{...} argument can be used to pass on parameters
to the \code{stat} part of the layer. An example of this is
\code{geom_area(stat = "density", adjust = 0.5)}. The stat's documentation
lists which parameters it can accept.
\item The \code{key_glyph} argument of \code{\link[ggplot2:layer]{layer()}} may also be passed on through
\code{...}. This can be one of the functions described as
\link[ggplot2:draw_key]{key glyphs}, to change the display of the layer in the legend.
}}

\item{bins}{Number of contour bins. Overridden by \code{breaks}.}

\item{binwidth}{The width of the contour bins. Overridden by \code{bins}.}

\item{breaks}{One of:
\itemize{
\item Numeric vector to set the contour breaks
\item A function that takes the range of the data and binwidth as input
and returns breaks as output. A function can be created from a formula
(e.g. ~ fullseq(.x, .y)).
}

Overrides \code{binwidth} and \code{bins}. By default, this is a vector of length
ten with \code{\link[=pretty]{pretty()}} breaks.}

\item{lineend}{Line end style (round, butt, square).}

\item{linejoin}{Line join style (round, mitre, bevel).}

\item{linemitre}{Line mitre limit (number greater than 1).}

\item{upscale_factor}{An integer by which to upscale the data before
computing the contours. For example if this is 2, the upscaled grid will
have 2 pixels in each direction from the grid before upscaling.}

\item{upscale_method}{The method used for upscaling. See
\code{\link[harpCore]{geo_upscale}} for more details.}

\item{downsample_location}{When "downsample" is the chosen method, each
pixel in the upscaled field is sampled from a pixel from the original field
that is inside the upscaled pixel. The location of that pixel can be one of
"bottom_left", "bottom_centre", "bottom_right", "left_centre", "centre",
"right_centre", "top_right", "top_centre", "top_left" or "random".}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\description{
These functions visualise \code{geofield} data as contours in 2D. They are built
on top of \code{\link[ggplot2]{geom_contour}} and
\code{\link[ggplot2]{geom_contour_filled}}.

They require data frames with a column containing
\code{\link[harpCore]{geofield}}s. This would normally be a
\code{\link[harpCore]{geolist}} column, but a standard list column containing
geofields will also work.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Smoothing can be achieved by upscaling
the data before plotting. This can only be done using an integer scale factor
and the methods can be any function that summarises a vector into a single
value, for example "mean", "median", "min", "max", or "downsample".
Downsampling is simply sampling a pixel within the upscaled pixel - by
default this is the centre.
}
\section{Aesthetics}{
 \code{geom_geocontour()} understands the following aesthetics
(required aesthetics are in bold):
\itemize{
\item \strong{geofield}
\item colour
\item alpha
\item group
\item linetype
\item linewidth
\item weight
}

\code{geom_geocontour_filled()} understands the following aesthetics (required
aesthetics are in bold):
\itemize{
\item \strong{geofield}
\item colour
\item fill
\item alpha
\item group
\item linetype
\item linewidth
\item weight
}
}

\section{Computed variables}{


These are calculated by the 'stat' part of layers and can be accessed with \link[ggplot2:aes_eval]{delayed evaluation}. The computed variables differ somewhat for contour lines  (computed by \code{stat_contour()}) and contour bands (filled contours,  computed by \code{stat_contour_filled()}). The variables \code{nlevel} and \code{piece}  are available for both, whereas \code{level_low}, \code{level_high}, and \code{level_mid}  are only available for bands. The variable \code{level} is a numeric or a factor  depending on whether lines or bands are calculated.
\itemize{
\item \code{after_stat(level)}\cr Height of contour. For contour lines, this is a numeric vector  that represents bin boundaries. For contour bands, this is an ordered  factor that represents bin ranges.
\item \code{after_stat(level_low)}, \code{after_stat(level_high)}, \code{after_stat(level_mid)}\cr (contour bands only) Lower and upper  bin boundaries for each band, as well as the mid point between boundaries.
\item \code{after_stat(nlevel)}\cr Height of contour, scaled to a maximum of 1.
\item \code{after_stat(piece)}\cr Contour piece (an integer).
}

}

