% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-geowindvec.R
\name{geom_geowindvec}
\alias{geom_geowindvec}
\title{Plot 2d wind vectors}
\usage{
geom_geowindvec(
  mapping = NULL,
  data = NULL,
  position = "identity",
  ...,
  rotate = FALSE,
  skip = 0,
  max_vec = 10,
  skip_from = "centre",
  arrow = ggplot2::arrow(length = unit(0.15, "cm")),
  arrow.fill = NULL,
  lineend = "butt",
  linejoin = "round",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{position}{Position adjustment, either as a string naming the adjustment
(e.g. \code{"jitter"} to use \code{position_jitter}), or the result of a call to a
position adjustment function. Use the latter if you need to change the
settings of the adjustment.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{rotate}{Whether to rotate the U and V wind components from grid
relative directions to earth relative directions prior to plotting}

\item{skip}{The number of grid squares to skip between each vector.}

\item{max_vec}{The magnitude of a vector with a length equal to \code{skip + 1}
grid squares.}

\item{skip_from}{The location of the grid square within a neighborhood of
skipped grid squares.}

\item{arrow}{specification for arrow heads, as created by \code{\link[grid:arrow]{grid::arrow()}}.}

\item{arrow.fill}{fill colour to use for the arrow head (if closed). \code{NULL}
means use \code{colour} aesthetic.}

\item{lineend}{Line end style (round, butt, square).}

\item{linejoin}{Line join style (round, mitre, bevel).}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\description{
This function visualises \code{geofield} \emph{U} and \emph{V} wind data as vectors. It is
built on top of \code{\link[ggplot2]{geom_segment}}.

It requires data frames with columns containing
\code{\link[harpCore]{geofield}}s for both the u and v winds. This would
normally be a \code{\link[harpCore]{geolist}} column, but a standard list
column containing geofields will also work.

Only one row of a data frame can be plotted without faceting.
}
\section{Aesthetics}{

\code{geom_geowindvec()} understands the following aesthetics (required aesthetics are
in bold):
\itemize{
\item \code{\strong{u}}
\item \code{\strong{v}}
\item \code{\link[ggplot2:aes_colour_fill_alpha]{alpha}}
\item \code{\link[ggplot2:aes_colour_fill_alpha]{colour}}
\item \code{\link[ggplot2:aes_group_order]{group}}
\item \code{\link[ggplot2:aes_linetype_size_shape]{linetype}}
\item \code{\link[ggplot2:aes_colour_fill_alpha]{linewidth}}
}
}

