% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_field.R
\name{plot_field}
\alias{plot_field}
\title{Plot a harp_fcst 2d field}
\usage{
plot_field(
  .fcst,
  fcst_model,
  plot_col,
  fcdate,
  lead_time,
  filter_by = NULL,
  palette = viridis::viridis(255),
  num_breaks = 15,
  breaks = NULL,
  legend = TRUE,
  title = "auto",
  zoom_centre = NULL,
  zoom_radius = 100
)
}
\arguments{
\item{.fcst}{A harp_fcst list or harp_spatial_fcst data frame.}

\item{fcst_model}{The name of the model to plot. This selects the model from
the harp_fcst list to plot. Can be omitted if there is only one model in
the harp_fcst list. For plotting directly from a harp_spatial_fcst data
frame, as returned by \code{\link{read_grid}}, this is the name to be used
in the plot title.}

\item{plot_col}{The column of data to plot from. Should be unquoted, or if a
variable, wrapped in double curly brackets: \{\{\}\}.}

\item{fcdate}{The forecast date to plot in "YYYYMMDDhh" or similar format.
Can be omitted if there is only one date in the data.}

\item{lead_time}{The lead time to plot. Can be omitted if there is only one
lead time in the data.}

\item{filter_by}{Expressions that return a logical value wrapped inside the
\code{vars} function for filtering the data prior to plotting. Only a
single row should be left in the data frame for \code{plot_field} to work.}

\item{palette}{Colour palette to use. This should be a vector of colours.}

\item{num_breaks}{Number of colour breaks to use in the plot.}

\item{breaks}{The values to use for colour breaks. If not NULL, breaks has
priority over num_breaks.}

\item{legend}{Logical. Whether to plot a legend.}

\item{title}{The title of the plot. Set to "auto" to automatically generate a
title.}

\item{zoom_centre}{The centre point for a zoomed plot. Should be a 2-element
vector with the longitude and latitude of the desired centre point.}

\item{zoom_radius}{The number of grid squares in each direction from
\code{zoom_centre} for a zoomed in plot.}
}
\value{
A plot
}
\description{
Plot a harp_fcst 2d field
}
