% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fss.R
\name{plot_fss}
\alias{plot_fss}
\title{Plot FSS scores}
\usage{
plot_fss(
  plot_data,
  score_name = "FSS",
  flip_axes = FALSE,
  colour_by = "fss",
  point_size = 1.2,
  line_width = 1,
  ...
)
}
\description{
\code{plot_spatial_verif} is used to plot verification scores computed by
functions from the harpSpatial package.
}
