% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_point_verif.R
\name{plot_point_verif}
\alias{plot_point_verif}
\title{Plot verification scores}
\usage{
plot_point_verif(verif_data, score, x_axis = leadtime,
  y_axis = rlang::enquo(score), colour_by = mname, facet_by = NULL,
  linetype_by = NULL, filter_by = NULL)
}
\arguments{
\item{verif_data}{Output from \link[harpPoint]{ens_verif},
\link[harpPoint]{det_verif}, or a \link[harpPoint] verification function
for individual scores.}

\item{score}{The score to plot. Should be the name of one of the columns in
the verification tables or the name of a dervived score, such as
\code{spread_skill}, \code{spread_skill_ratio}, or
\code{brier_score_decomposition}.}

\item{x_axis}{The x-axis for the plot. The default is leadtime, but could
also be threshold. For some scores this is overrided.}

\item{y_axis}{The y-axis for the plot. The default is to take the same as the
score input, and for most scores this is overrided.}

\item{colour_by}{The column to colour the plot lines or bars by. The default
is mname, for the model name. Set to NULL for all lines / bars to have the
same colour.}

\item{facet_by}{The column(s) to facet the plot by. Faceting is a term used
for generating plot panels. The argument must be wrapped inside the
\code[dplyr]{vars} function - e.g. \code{facet_by = vars(threshold)}.}

\item{linetype_by}{The column to set the line types of the plot by.}

\item{filter_by}{Filter the data before plotting. Must be wrapped inside the
\code[dplyr]{vars} function. This can be useful for making a single plot
where there are many groups. For example, for reliability there should be
one plot for each lead time and threshold, so the data can be filtered with
e.g. \code{filter_by = vars(leadtime == 12, threshold == 280)}.}
}
\value{
A plot. Can be saved with \link[ggplot2]{ggsave}.
}
\description{
\code{plot_point_verif} is used to plot verification scores computed by
functions from the harpPoint package. The function uses non standard
evaluation (NSE) meaning that none of the arguments should be quoted. For
arguments that can take more than one value, (\code{facet_by} and
\code{filter_by}), the arguments should be wrapped inside the
\code[dplyr]{vars} function.
}
\examples{
plot_point_verif(ens_verif_data, crps)
plot_point_verif(ens_verif_data, spread_skill)
plot_point_verif(det_verif_data, equitable_threat_score, facet_by = vars(threshold))
plot_point_verif(ens_verif_data, reliability, filter_by = vars(leadtime == 12, threshold == 280))
}

