% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sal.R
\name{plot_sal}
\alias{plot_sal}
\title{Plot SAL scores}
\usage{
plot_sal(
  plot_data,
  score_name = "SAL",
  point_size = 4,
  extend_y_to_zero = TRUE,
  line_width = 1,
  line_colour = "grey80",
  quants = seq(0.25, 0.75, 0.05),
  xylim = 2,
  ...
)
}
\description{
\code{plot_spatial_verif} is used to plot verification scores computed by
functions from the harpSpatial package.
}
