% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_station_ts.R
\name{plot_station_ts}
\alias{plot_station_ts}
\alias{plot_station_ts.harp_ens_point_df}
\title{Plot a time series for point data}
\usage{
plot_station_ts(
  .data,
  SID,
  fcst_dttm,
  x_axis = "lead_time",
  fcst_geom = "line",
  fcst_geom_args = list(),
  fcst_colour_by = NULL,
  fcst_colours = NULL,
  obs_col = NULL,
  obs_geom = "point",
  obs_geom_args = list(),
  facet_by = NULL,
  num_facet_cols = 1,
  facet_scales = "free_y",
  smooth = FALSE,
  ...
)

\method{plot_station_ts}{harp_ens_point_df}(
  .data,
  SID,
  fcst_dttm,
  x_axis = "lead_time",
  fcst_geom = "boxplot",
  fcst_geom_args = list(),
  fcst_colour_by = NULL,
  fcst_colours = NULL,
  obs_col = NULL,
  obs_geom = "point",
  obs_geom_args = list(),
  facet_by = NULL,
  num_facet_cols = 1,
  facet_scales = "free_y",
  smooth = FALSE,
  quantiles = NULL,
  best_guess = NULL,
  best_guess_geom = "line",
  best_guess_geom_args = list(),
  ...
)
}
\arguments{
\item{.data}{A \code{harp_point_df} data frame.}

\item{SID}{The ID of the station(s) to plot. If more than one SID is asked
for then \code{SID} should be included in \code{facet_by}.}

\item{fcst_dttm}{The start time(s) of the the forecast to plot. If more than
one \code{fcst_dttm} is asked for, \code{fcst_dttm} should be included in \code{facet_by}.}

\item{x_axis}{The x axis of the plot.}

\item{fcst_geom}{The geom to use to plot the forecast data (see details).}

\item{fcst_geom_args}{Arguments to the \code{fcst_geom} geom as a named list.}

\item{fcst_colour_by}{Which column in \code{.data} to use to control the colours
of the forecast data.}

\item{fcst_colours}{A vector of colours to use for the forecast data. It
should be the same length as the number of colours to appear in the plot.}

\item{obs_col}{If observations are to be included in the plot, the column
containing the observations data.}

\item{obs_geom}{The geom to use to plot the observations data.}

\item{obs_geom_args}{Arguments to the \code{obs_geom} geom as a named list.}

\item{facet_by}{The column(s) to use to facet the plot into panels.}

\item{num_facet_cols}{The number of columns in a faceted plot.}

\item{facet_scales}{Whether the scales should be fixed. Defaults to
\code{"free_y"}. See \code{\link[ggplot2]{facet_wrap}} for more details.}

\item{smooth}{For line and ribbon plots, whether to smooth the lines by
drawing an X-spline relative to control points. In the background
\code{\link{geom_linespline}} and \code{\link{geom_ribbonspline}} are used.}

\item{...}{Other arguments passed to methods.}

\item{quantiles}{For \code{geom = "ribbon"}, or \code{geom = "col"}, the quantiles
used to stratify the probabilities of an ensemble forecast.}

\item{best_guess}{What to plot as a "best guess" forecast. Can be any
function as a character string that reduces a vector to a single value. Can
also be an ensemble member as a numeric value or a string that is the same
as the member in a \code{harp_ens_point_df} data frame that has had the
members pivoted using \code{\link[harpCore]{pivot_members}}, e.g.
\code{"mbr000"}.}

\item{best_guess_geom}{The geom to use to plot the best guess forecast.}

\item{best_guess_geom_args}{Arguments to \code{best_guess_geom} as a named list.}
}
\value{
A ggplot object that can be saved with \code{\link[ggplot2]{ggasve}}.
}
\description{
Plot a time series for point data
}
