% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vertical_profile.R
\name{plot_vertical_profile}
\alias{plot_vertical_profile}
\title{Title}
\usage{
plot_vertical_profile(.fcst, SID, fcdate, lead_time, y_axis = p,
  skew_t = FALSE, colour_by = parameter, colours = NULL,
  facet_by = NULL, num_facet_cols = 2, ...)
}
\arguments{
\item{colours}{}
}
\description{
Title
}
