% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_plot_point_verif.R
\name{shiny_plot_point_verif}
\alias{shiny_plot_point_verif}
\title{Start shiny app to plot point verification}
\usage{
shiny_plot_point_verif(
  start_dir = NULL,
  theme = c("dark", "light", "white"),
  online = TRUE,
  ...
)
}
\description{
Start shiny app to plot point verification
}
